Added a readme and a way to test basic shape rendering.

+++ src/lights.cpp
@@ -2,9 +2,9 @@
 #include "mazeGen.hpp"
 #include "draw.hpp"
 
-static GLfloat getCurrentLight(int i)
+static GLint getCurrentLight(int i)
 {
-    GLfloat curLight;
+    GLint curLight;
     switch (i)
     {
     case 0:
